close to goal

+++ .laborantum/texts/Homeworks/1. Vectors/17. Elementwise Arithmetics Exercise/task.py
@@ -2,5 +2,5 @@ import numpy as np
 
 answer = {
     'task1': np.array([7.5, 6.5, 4.25, 5.75]),
-    'task2': np.array([2.,0.,0.,0.])
+    'task2': np.array([2.,0.,0.,0.]).astype("int64")
 }
